Code/requirements.txt
openai
python-dotenv
requests
pandas
scipy
random 
numpy